Process validation correctly: returning 400 error instead of 422; refactor api structure
