+++ .pre-commit-config.yaml
@@ -9,17 +9,21 @@ repos:
       - id: check-yaml
         args:
           - --unsafe
-      - id: end-of-file-fixer
-        exclude: ^frontend/src/client/.*
-      - id: trailing-whitespace
-        exclude: ^frontend/src/client/.*
-  - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.2.2
+  - repo: https://github.com/psf/black
+    rev: 24.4.2
     hooks:
-      - id: ruff
+      - id: black
+        args: [ "--line-length=79", "--target-version=py311" ]
+  - repo: https://github.com/PyCQA/isort
+    rev: 5.13.2
+    hooks:
+      - id: isort
         args:
-          - --fix
-      - id: ruff-format
+          - --profile=black
+          - --line-length=79
+          - --combine-as
+          - --combine-star
+          - --remove-redundant-aliases
 
 ci:
   autofix_commit_msg: 🎨 [pre-commit.ci] Auto format from pre-commit.com hooks
