+++ main.py
@@ -1,5 +1,7 @@
 import sentry_sdk
-from fastapi import FastAPI
+from fastapi import FastAPI, Request, status
+from fastapi.exceptions import RequestValidationError
+from fastapi.responses import JSONResponse
 from fastapi.routing import APIRoute
 from starlette.middleware.cors import CORSMiddleware
 
@@ -19,6 +21,25 @@ app = FastAPI(
     openapi_url=f"{settings.API_V1_STR}/openapi.json",
     generate_unique_id_function=custom_generate_unique_id,
 )
+
+
+@app.exception_handler(RequestValidationError)
+async def validation_exception_handler(
+    request: Request, exc: RequestValidationError
+):
+    errors: list[dict[str, str]] = [
+        {
+            "loc": " -> ".join(str(loc) for loc in err["loc"]),
+            "msg": err["msg"],
+            "type": err["type"],
+        }
+        for err in exc.errors()
+    ]
+    return JSONResponse(
+        status_code=status.HTTP_400_BAD_REQUEST,
+        content={"detail": errors},
+    )
+
 
 # Set all CORS enabled origins
 if settings.BACKEND_CORS_ORIGINS:
