+++ requirements.txt
@@ -1,56 +1,58 @@
-alembic==1.13.1 ; python_version >= "3.10" and python_version < "4.0"
-annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
-anyio==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
-bcrypt==4.0.1 ; python_version >= "3.10" and python_version < "4.0"
-cachetools==5.3.3 ; python_version >= "3.10" and python_version < "4.0"
-certifi==2024.2.2 ; python_version >= "3.10" and python_version < "4.0"
-chardet==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
-charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
-click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
-colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
-cssselect==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
-cssutils==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
-dnspython==2.6.1 ; python_version >= "3.10" and python_version < "4.0"
-email-validator==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
-emails==0.6 ; python_version >= "3.10" and python_version < "4.0"
-exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
-fastapi==0.109.2 ; python_version >= "3.10" and python_version < "4.0"
-greenlet==3.0.3 ; python_version >= "3.10" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
-gunicorn==22.0.0 ; python_version >= "3.10" and python_version < "4.0"
-h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
-httpcore==1.0.4 ; python_version >= "3.10" and python_version < "4.0"
-httptools==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
-httpx==0.25.2 ; python_version >= "3.10" and python_version < "4.0"
-idna==3.7 ; python_version >= "3.10" and python_version < "4.0"
-jinja2==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
-lxml==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
-mako==1.3.2 ; python_version >= "3.10" and python_version < "4.0"
-markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
-packaging==24.0 ; python_version >= "3.10" and python_version < "4.0"
-passlib[bcrypt]==1.7.4 ; python_version >= "3.10" and python_version < "4.0"
-premailer==3.10.0 ; python_version >= "3.10" and python_version < "4.0"
-psycopg-binary==3.1.18 ; implementation_name != "pypy" and python_version >= "3.10" and python_version < "4.0"
-psycopg[binary]==3.1.18 ; python_version >= "3.10" and python_version < "4.0"
-pydantic-core==2.16.3 ; python_version >= "3.10" and python_version < "4.0"
-pydantic-settings==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
-pydantic==2.6.4 ; python_version >= "3.10" and python_version < "4.0"
-pyjwt==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
-python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
-python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
-python-multipart==0.0.7 ; python_version >= "3.10" and python_version < "4.0"
-pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
-requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
-sentry-sdk[fastapi]==1.41.0 ; python_version >= "3.10" and python_version < "4.0"
-six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
-sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
-sqlalchemy==2.0.28 ; python_version >= "3.10" and python_version < "4.0"
-sqlmodel==0.0.16 ; python_version >= "3.10" and python_version < "4.0"
-starlette==0.36.3 ; python_version >= "3.10" and python_version < "4.0"
-tenacity==8.2.3 ; python_version >= "3.10" and python_version < "4.0"
-typing-extensions==4.10.0 ; python_version >= "3.10" and python_version < "4.0"
-tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
-urllib3==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
-uvicorn[standard]==0.24.0.post1 ; python_version >= "3.10" and python_version < "4.0"
-uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
-watchfiles==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
-websockets==12.0 ; python_version >= "3.10" and python_version < "4.0"
+alembic==1.13.1
+annotated-types==0.6.0
+anyio==4.3.0
+bcrypt==4.0.1
+cachetools==5.3.3
+certifi==2024.2.2
+chardet==5.2.0
+charset-normalizer==3.3.2
+click==8.1.7
+colorama==0.4.6
+cssselect==1.2.0
+cssutils==2.9.0
+dnspython==2.6.1
+email-validator==2.1.1
+emails==0.6
+exceptiongroup==1.2.0
+fastapi==0.109.2
+greenlet==3.0.3
+gunicorn==22.0.0
+h11==0.14.0
+httpcore==1.0.4
+httptools==0.6.1
+httpx==0.25.2
+idna==3.7
+jinja2==3.1.4
+lxml==5.1.0
+mako==1.3.2
+markupsafe==2.1.5
+packaging==24.0
+passlib[bcrypt]==1.7.4
+premailer==3.10.0
+psycopg-binary==3.1.18
+psycopg[binary]==3.1.18
+pydantic-core==2.16.3
+pydantic-settings==2.2.1
+pydantic==2.6.4
+pyjwt==2.8.0
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.1
+python-multipart==0.0.7
+pyyaml==6.0.1
+requests==2.31.0
+sentry-sdk[fastapi]==1.41.0
+six==1.16.0
+sniffio==1.3.1
+sqlalchemy==2.0.28
+sqlmodel==0.0.16
+starlette==0.36.3
+tenacity==8.2.3
+typing-extensions==4.10.0
+tzdata==2024.1
+urllib3==2.2.1
+uvicorn[standard]==0.24.0.post1
+uvloop==0.19.0
+watchfiles==0.21.0
+websockets==12.0
+black==24.4.2
+isort==5.13.2
